Remove the buttons from the landing page
